requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
PyJWT==2.8.0
pandas==1.5.3
numpy==1.26.4
joblib==1.3.2
scikit-learn==1.3.0
openpyxl==3.1.2
matplotlib==3.8.0
gunicorn==21.2.0
